layout fixes

+++ components/Item.js
@@ -1,5 +1,5 @@
 import React, { useEffect } from 'react';
-import { Text, StyleSheet, View , Pressable, Alert } from 'react-native';
+import { Text, StyleSheet, View, Pressable, Alert } from 'react-native';
 import { PanGestureHandler } from 'react-native-gesture-handler';
 import Animated, { withSpring, useAnimatedGestureHandler, useAnimatedStyle, useSharedValue, runOnJS } from 'react-native-reanimated';
 import MaterialCommunityIcons from 'react-native-vector-icons/MaterialCommunityIcons';
@@ -7,18 +7,19 @@ import { useTheme } from 'react-native-paper';
 
 const SWIPE_WIDTH = 60;
 
-const Item = ({ label, active, setOnActive, index, deleteItem}) => {
+const Item = ({ label, active, setOnActive, index, deleteItem, createdAt }) => {
 
     const theme = useTheme();
     const translateX = useSharedValue(-300);
 
     useEffect(() => {
-        console.log("active changed to ", active);
+
         (active !== index) && closeTab();
     }, [active]);
 
     useEffect(() => {
         translateX.value = withSpring(0);
+
     }, []);
 
     const animatedStyle = useAnimatedStyle(() => {
@@ -30,21 +31,35 @@ const Item = ({ label, active, setOnActive, index, deleteItem}) => {
     });
 
     const onPressHandler = () => {
-        Alert.alert("Delete Item? " , 
-        "Your goal will be deleted" , 
-        [ {text: "OK", style : "destructive", onPress: deleteItem,}  
-        , {text: "Cancel", onPress: ()=> {closeTab()}}
-    ] );
+        Alert.alert("Delete Item? ",
+            "Your goal will be deleted",
+            [{ text: "OK", style: "destructive", onPress: deleteItem, }
+                , { text: "Cancel", onPress: () => { closeTab() } }
+            ]);
     };
+
+    const CompletedHandler = () => {
+        Alert.alert("Completed Item? ",
+            "Your goal will be deleted",
+            [{ text: "OK", style: "default", onPress: deleteItem, }
+                , { text: "Cancel", onPress: () => { closeTab() } }
+            ]);
+
+    }
 
     const closeTab = () => {
         'worklet';
         translateX.value = withSpring(0);
     }
 
-    const openTab = () => {
+    const openTabRight = () => {
         'worklet';
         translateX.value = withSpring(-SWIPE_WIDTH);
+        runOnJS(setOnActive)(index);
+    }
+    const openTabLeft = () => {
+        'worklet';
+        translateX.value = withSpring(SWIPE_WIDTH);
         runOnJS(setOnActive)(index);
     }
 
@@ -58,7 +73,10 @@ const Item = ({ label, active, setOnActive, index, deleteItem}) => {
         },
         onEnd: () => {
             if (translateX.value < -SWIPE_WIDTH / 2) {
-                openTab();
+                openTabRight();
+            }
+            else if (translateX.value > SWIPE_WIDTH / 2 + 4) {
+                openTabLeft();
             }
             else {
                 closeTab();
@@ -66,19 +84,36 @@ const Item = ({ label, active, setOnActive, index, deleteItem}) => {
         },
     });
 
+
     return (
         <View>
-            <PanGestureHandler onGestureEvent={onGestureEvent}>
-                <Animated.View style={[styles.container, animatedStyle, {backgroundColor: theme.colors.card, borderColor: theme.colors.border}]}>
-                    <Text style={{...styles.label, color: theme.colors.text}}>{label}</Text>
+            <PanGestureHandler   activeOffsetX={[-1,1]} onGestureEvent={onGestureEvent}>
+                <Animated.View style={[styles.container, animatedStyle, { backgroundColor: theme.colors.card, borderColor: theme.colors.border }]}>
+                    <Text style={styles.hourStyle}>
+                        {[createdAt.getHours(), ":", createdAt.
+                            getMinutes().toString().length === 1 ? ("0" + createdAt.getMinutes()) : createdAt.getMinutes()].join(" ")}</Text>
+                    <Text 
+                    numberOfLines={1}
+                    ellipsizeMode="tail"
+                    style={{ ...styles.labelStyle, color: theme.colors.text }}>{label}</Text>
+                    <Text style={ styles.dateStyle }>{createdAt.toLocaleDateString()}</Text>
                 </Animated.View>
             </PanGestureHandler>
-            <View style={styles.swipeContainer}>
-                <Pressable 
-                onPress={onPressHandler}
+            <View style={styles.leftSwipeContainer}>
+                <Pressable
+                    onPress={onPressHandler}
                 >
-                <MaterialCommunityIcons name="trash-can" size={30} color="#fff" />
+                    <MaterialCommunityIcons name="trash-can" size={30} color="#fff" />
                 </Pressable>
+
+            </View>
+            <View style={styles.rightSwipeContainer}>
+                <Pressable
+                    onPress={CompletedHandler}
+                >
+                    <MaterialCommunityIcons name="check" size={30} color="#fff" />
+                </Pressable>
+
             </View>
         </View>
     )
@@ -89,16 +124,18 @@ export default Item;
 
 const styles = StyleSheet.create({
     container: {
+
         height: 60,
         alignItems: 'center',
         zIndex: 1,
         justifyContent: 'center',
         width: '100%',
-        overflow: 'hidden',
         borderBottomWidth: 1,
+        borderTopWidth: 1,
         borderColor: '#8e8e93'
     },
-    swipeContainer: {
+    leftSwipeContainer: {
+
         position: 'absolute',
         zIndex: -1,
         right: 0,
@@ -108,4 +145,46 @@ const styles = StyleSheet.create({
         justifyContent: 'center',
         backgroundColor: 'red',
     },
+
+    labelStyle: {
+        position: 'absolute',
+        fontSize: 15,
+        fontStyle: 'italic',
+        fontWeight: 'bold',
+        maxWidth: '60%',
+        textAlign: 'left', 
+        left:'10%',
+    },
+
+    rightSwipeContainer: {
+
+        position: 'absolute',
+        zIndex: -1,
+        left: 0,
+        height: 60,
+        width: SWIPE_WIDTH,
+        alignItems: 'center',
+        justifyContent: 'center',
+        backgroundColor: 'green',
+    },
+
+    dateStyle: {
+        fontSize: 12,
+        color: 'grey',
+        position: 'absolute',
+        right: 10,
+        bottom: 15,
+    },
+
+    hourStyle: {
+        fontSize: 12,
+        color: 'grey',
+        position: 'absolute',
+        right: 10,
+        bottom: 35,
+    },
+
+
+
+
 });
